# api/docs/swagger.yaml
definitions:
...
  github_com_time_capsule_Api-Gateway-Timecapsule_models.CreateHistoricalEventModel:
    properties:
      category:
        type: string
      date:
        type: string
      description:
        type: string
      id:
        type: string
      source_url:
        type: string
      title:
        type: string
      user_id:
        description: Added user_id field
        type: string
    type: object
...
  github_com_time_capsule_Api-Gateway-Timecapsule_models.PatchHistoricalEventModel:
    properties:
      category:
        type: string
      date:
        type: string
      description:
        type: string
      id:
        type: string
      source_url:
        type: string
      title:
        type: string
      user_id:
        description: Added user_id field (optional)
        type: string
    type: object
...
  github_com_time_capsule_Api-Gateway-Timecapsule_models.UpdateHistoricalEventModel:
    properties:
      category:
        type: string
      date:
        type: string
      description:
        type: string
      id:
        type: string
      source_url:
        type: string
      title:
        type: string
      user_id:
        description: Added user_id field
        type: string
    type: object
...
  timeline.GetAllHistoricalEventsResponse:
    properties:
      count:
        type: integer
      historical_events:
        items:
          $ref: '#/definitions/timeline.HistoricalEvent'
        type: array
    type: object
...
  timeline.HistoricalEvent:
    properties:
      category:
        type: string
      created_at:
        type: string
      date:
        description: Use string for date (YYYY-MM-DD format)
        type: string
      description:
        type: string
      id:
        type: string
      source_url:
        type: string
      title:
        type: string
      user_id:
        description: Added user_id field
        type: string
    type: object
...
info:
  contact: {}
  description: This is a sample server celler server.
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
...
  /v1/historical-events:
    get:
      consumes:
      - application/json
      description: Get a list of historical events.
      parameters:
      - description: Page number
        in: query
        minimum: 1
        name: page
        type: integer
      - description: Number of items per page
        in: query
        minimum: 1
        name: limit
        type: integer
      - description: Filter by title
        in: query
        name: title
        type: string
      - description: Filter by category
        in: query
        name: category
        type: string
      - description: Filter by start date (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: Filter by end date (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/timeline.GetAllHistoricalEventsResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: List Historical Events
      tags:
      - HistoricalEvents
    post:
      consumes:
      - application/json
      description: Create a new historical event with the provided details.
      parameters:
      - description: Historical Event details
        in: body
        name: historicalEvent
        required: true
        schema:
          $ref: '#/definitions/github_com_time_capsule_Api-Gateway-Timecapsule_models.CreateHistoricalEventModel'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create a new Historical Event
      tags:
      - HistoricalEvents
  /v1/historical-events/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a historical event by its ID.
      parameters:
      - description: Historical Event ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete Historical Event
      tags:
      - HistoricalEvents
    get:
      consumes:
      - application/json
      description: Get details of a historical event by its ID.
      parameters:
      - description: Historical Event ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/timeline.HistoricalEvent'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get Historical Event by ID
      tags:
      - HistoricalEvents
    patch:
      consumes:
      - application/json
      description: Partially update an existing historical event.
      parameters:
      - description: Historical Event ID
        in: path
        name: id
        required: true
        type: string
      - description: Patched historical event data
        in: body
        name: historicalEvent
        required: true
        schema:
          $ref: '#/definitions/github_com_time_capsule_Api-Gateway-Timecapsule_models.PatchHistoricalEventModel'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Partially update Historical Event
      tags:
      - HistoricalEvents
    put:
      consumes:
      - application/json
      description: Update an existing historical event.
      parameters:
      - description: Historical Event ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated historical event data
        in: body
        name: historicalEvent
        required: true
        schema:
          $ref: '#/definitions/github_com_time_capsule_Api-Gateway-Timecapsule_models.UpdateHistoricalEventModel'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update Historical Event
      tags:
      - HistoricalEvents
...
securityDefinitions:
  ApiKeyAuth:
    description: Description for what is this security definition being used
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"